Time: 28 ms (64.52%) | Memory: 27.1 MB (52.06%) - LeetSync

+++ 115-distinct-subsequences/distinct-subsequences.cpp
@@ -22,7 +22,7 @@ public:
         
         return dp[i][j];
     }
-    long long solveTab(string &s, string &t) {
+    int solveTab(string &s, string &t) {
         vector<vector<int>> dp(s.size() + 1, vector<int>(t.size() + 1, 0));
 
         for (int i = 0; i <= s.size(); i++) {
